requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.3
qrcode[pil]==7.4.2
Pillow==10.4.0
python-dotenv==1.0.1
twilio==9.3.3
gunicorn==22.0.0
